projet/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.29.5
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.335
langsmith==0.0.63
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.2
openai==0.28.0
opencv-python==4.8.1.78
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pdf2image==1.16.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
posthog==3.0.2
prompt-toolkit==3.0.40
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
PyPDF2==1.28.6
PyPika==0.48.9
pysqlite3-binary==0.5.2.post1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.1
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.20.0
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.9
websocket-client==1.6.4
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
